Organize project under src package
