+++ src/face_clustering/__init__.py
@@ -1,0 +1,18 @@
+"""Face clustering utilities package."""
+
+from .cluster_faces import cluster_face_embeddings
+from .crop_bodies import crop_person
+from .detect_bibs import detect_bib_in_crop
+from .detect_runners import detect_persons
+from .face_embeddings import extract_face_embeddings
+from .visualize_embeddings import plot_embeddings, reduce_embeddings
+
+__all__ = [
+    "cluster_face_embeddings",
+    "crop_person",
+    "detect_bib_in_crop",
+    "detect_persons",
+    "extract_face_embeddings",
+    "plot_embeddings",
+    "reduce_embeddings",
+]
