+++ src/face_clustering/crop_bodies.py
