+++ src/face_clustering/extract_faces.py
