+++ src/face_clustering/visualize_embeddings.py
